requirements.txt
flask
sentence-transformers
ddgs
requests
faiss-cpu
numpy
streamlit
pyperclip
